test(memory): colocate optimizeObservationsForContext tests with source
- Moves the optimizeObservationsForContext suite out of __tests__ into
  observer-agent.test.ts, beside the implementation it covers
- Drops the now-unused import from the observational-memory suite

Per AGENTS.md, vitest tests are colocated with source.
